backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
ifcopenshell
xmlschema
numpy
odfpy
pystache
bcf-client
python-dateutil
whitenoise>=6.5,<7.0
